minor changes in poc-client
had to make minor changes (due to a mistake)
