--- poc-client/app/about/page.tsx
@@ -1,7 +1,0 @@
-export default function About() {
-  return (
-    <main className="min-h-screen bg-white text-black">
-      <h1>About Stagnum</h1>
-    </main>
-  );
-}
