--- poc-client/app/contact/page.tsx
@@ -1,7 +1,0 @@
-export default function Contact() {
-  return (
-    <main className="min-h-screen bg-white text-black">
-      <h1>Contact</h1>
-    </main>
-  );
-}
