--- poc-client/poc-client/next.config.mjs
@@ -1,4 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {reactStrictMode: false};
-
-export default nextConfig;
